backend/requirements.txt
# Requirements for Python backend
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0

# AI and ML
google-generativeai==0.3.2

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
Jinja2==3.1.3
WeasyPrint==62.1
reportlab==4.0.7

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database (optional)
sqlalchemy==2.0.25
alembic==1.13.1

# Monitoring and logging
structlog==23.2.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.2
black==23.12.1
flake8==7.0.0